fix cut-n-paste bug that crashed LAMMPS on reading molecular data files with -DLAMMPS_BIGBIG

+++ src/atom_vec.cpp
@@ -237,7 +237,7 @@ void AtomVec::grow(int n)
         memory->grow(*((bigint ***) pdata),nmax*nthreads,cols,"atom:barray");
       else {
         maxcols = *(mgrow.maxcols[i]);
-        memory->grow(*((int ***) pdata),nmax*nthreads,maxcols,"atom:barray");
+        memory->grow(*((bigint ***) pdata),nmax*nthreads,maxcols,"atom:barray");
       }
     }
   }
